+++ src/docker/start.sh
@@ -32,6 +32,9 @@ sed -i "s/{icinga-password}/${ICINGA_PASSWORD}/"        /usr/local/tomcat/conf/s
 sed -i "s/{icinga-site-suffix}/${ICINGA_SITE_SUFFIX}/"  /usr/local/tomcat/conf/samply.share.broker.conf
 sed -i "s/{icinga-project}/${ICINGA_PROJECT}/"          /usr/local/tomcat/conf/samply.share.broker.conf
 
+sed -i "s/{structured-query-username}/${STRUCTURED_QUERY_USERNAME}/"          /usr/local/tomcat/conf/samply.share.broker.conf
+sed -i "s/{structured-query-password}/${STRUCTURED_QUERY_PASSWORD}/"          /usr/local/tomcat/conf/samply.share.broker.conf
+
 file=${CATALINA_HOME}/conf/features.properties
 sed -i "s/{FEATURE_CREATE_NEW_SITE}/${FEATURE_CREATE_NEW_SITE:-false}/" "$file"