+++ src/main/java/de/samply/share/broker/rest/InquiryHandler.java
@@ -1,75 +1,62 @@
 package de.samply.share.broker.rest;
 
+import static java.lang.Boolean.TRUE;
+import static javax.xml.bind.Marshaller.JAXB_FRAGMENT;
+
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import de.samply.share.broker.jdbc.ResourceManager;
+import de.samply.share.broker.model.DefaultInquiryCriteriaTranslatable;
+import de.samply.share.broker.model.InquiryCriteriaTranslatable;
 import de.samply.share.broker.model.db.Tables;
-import de.samply.share.broker.model.db.enums.ActionType;
 import de.samply.share.broker.model.db.enums.DocumentType;
-import de.samply.share.broker.model.db.enums.InquiryCriteriaType;
 import de.samply.share.broker.model.db.enums.InquiryStatus;
 import de.samply.share.broker.model.db.enums.ProjectStatus;
 import de.samply.share.broker.model.db.tables.daos.ContactDao;
 import de.samply.share.broker.model.db.tables.daos.InquiryDao;
 import de.samply.share.broker.model.db.tables.daos.ReplyDao;
 import de.samply.share.broker.model.db.tables.daos.UserDao;
-import de.samply.share.broker.model.db.tables.pojos.Action;
 import de.samply.share.broker.model.db.tables.pojos.BankSite;
 import de.samply.share.broker.model.db.tables.pojos.Document;
 import de.samply.share.broker.model.db.tables.pojos.Inquiry;
 import de.samply.share.broker.model.db.tables.pojos.InquiryCriteria;
-import de.samply.share.broker.model.db.tables.pojos.Project;
 import de.samply.share.broker.model.db.tables.pojos.Reply;
 import de.samply.share.broker.model.db.tables.pojos.Site;
 import de.samply.share.broker.model.db.tables.pojos.User;
-import de.samply.share.broker.statistics.StatisticsHandler;
-import de.samply.share.broker.utils.EssentialSimpleQueryDto2ShareXmlTransformer;
-import de.samply.share.broker.utils.cql.EssentialSimpleQueryDto2CqlTransformer;
-import de.samply.share.broker.utils.db.ActionUtil;
 import de.samply.share.broker.utils.db.BankSiteUtil;
 import de.samply.share.broker.utils.db.ContactUtil;
 import de.samply.share.broker.utils.db.DocumentUtil;
 import de.samply.share.broker.utils.db.InquiryCriteriaUtil;
 import de.samply.share.broker.utils.db.InquiryUtil;
-import de.samply.share.broker.utils.db.ProjectUtil;
 import de.samply.share.broker.utils.db.SiteUtil;
 import de.samply.share.common.utils.Constants;
 import de.samply.share.common.utils.ProjectInfo;
 import de.samply.share.common.utils.SamplyShareUtils;
-import de.samply.share.essentialquery.EssentialSimpleFieldDto;
-import de.samply.share.essentialquery.EssentialSimpleQueryDto;
-import de.samply.share.essentialquery.EssentialSimpleValueDto;
-import de.samply.share.model.common.And;
 import de.samply.share.model.common.Contact;
 import de.samply.share.model.common.Info;
 import de.samply.share.model.common.ObjectFactory;
 import de.samply.share.model.common.Query;
-import de.samply.share.model.common.Where;
 import de.samply.share.model.common.inquiry.InquiriesIdList;
 import de.samply.share.model.cql.CqlQuery;
 import de.samply.share.model.cql.CqlQueryList;
-import de.samply.share.query.enums.SimpleValueCondition;
 import java.io.File;
 import java.io.IOException;
 import java.io.StringReader;
 import java.io.StringWriter;
-import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
-import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -86,13 +73,9 @@ public class InquiryHandler {
 
   private static final Logger logger = LogManager.getLogger(InquiryHandler.class);
 
-  private static final String ENTITY_TYPE_FOR_QUERY = "Donor + Sample";
-  private static final String ENTITY_TYPE_FOR_CQL_PATIENT = "Patient";
-  private static final String ENTITY_TYPE_FOR_CQL_SPECIMEN = "Specimen";
-
   private static final String QUERYLANGUAGE_QUERY = "QUERY";
   private static final String QUERYLANGUAGE_CQL = "CQL";
-  private static final Gson gson = new Gson();
+  private static final String QUERYLANGUAGE_STRUCTURED_QUERY = "STRUCTURED_QUERY";
 
   public InquiryHandler() {
   }
@@ -105,11 +88,15 @@ public class InquiryHandler {
     return StringUtils.equalsIgnoreCase(queryLanguage, QUERYLANGUAGE_CQL);
   }
 
+  static boolean isQueryLanguageStructuredQuery(String queryLanguage) {
+    return StringUtils.equalsIgnoreCase(queryLanguage, QUERYLANGUAGE_STRUCTURED_QUERY);
+  }
+
   /**
    * Store an inquiry and release it (or wait for ccp office authorization first).
    *
-   * @param simpleQueryDtoJson the query criteria of the inquiry as EssentialSimpleQueryDto
-   *                           represented as XML
+   * @param query              a representation of the query which can be translated to inquiry
+   *                           criteria
    * @param userid             the id of the user that releases the inquiry
    * @param inquiryName        the label of the inquiry
    * @param inquiryDescription the description of the inquiry
@@ -117,60 +104,35 @@ public class InquiryHandler {
    * @param voteId             the id of the vote linked with this inquiry
    * @param resultTypes        list of the entities that are searched for
    * @param bypassExamination  if true, no check by the ccp office is required
-   * @param isMonitoring       if true, no creation of cql query
    * @return the id of the inquiry
    */
-  public int storeAndRelease(String simpleQueryDtoJson, int userid, String inquiryName,
+  public int storeAndRelease(InquiryCriteriaTranslatable query,
+      int userid, String inquiryName,
       String inquiryDescription, int exposeId, int voteId, List<String> resultTypes,
-      boolean bypassExamination, boolean isMonitoring) {
-    int inquiryId = store(simpleQueryDtoJson, userid, inquiryName, inquiryDescription,
-        exposeId, voteId, resultTypes, isMonitoring);
+      boolean bypassExamination) throws IllegalStateException {
+    int inquiryId = store(query, userid, inquiryName, inquiryDescription,
+        exposeId, voteId, resultTypes);
     Inquiry inquiry = InquiryUtil.fetchInquiryById(inquiryId);
     release(inquiry, bypassExamination);
     return inquiryId;
   }
 
-  private EssentialSimpleQueryDto jsonString2EssentialDto(String simpleQueryDtoJson) {
-    EssentialSimpleQueryDto queryDto = new EssentialSimpleQueryDto();
-
-    try {
-      Gson gson = new Gson();
-      queryDto = gson.fromJson(simpleQueryDtoJson, EssentialSimpleQueryDto.class);
-    } catch (Exception e) {
-      logger.error(e);
-      e.printStackTrace();
-    }
-
-    for (EssentialSimpleFieldDto field : queryDto.getFieldDtos()) {
-      field.setValueDtos(
-          field.getValueDtos().stream()
-              .filter(valueDto -> !isEmpty(valueDto))
-              .collect(Collectors.toList()));
-    }
-
-    queryDto.setFieldDtos(queryDto.getFieldDtos().stream()
-        .filter(fieldDto -> !isEmpty(fieldDto))
-        .collect(Collectors.toList()));
-
-    return queryDto;
-  }
-
   /**
    * Store a new inquiry draft.
    *
-   * @param query              the query criteria of the inquiry
+   * @param query              a representation of the query which can be translated to inquiry
+   *                           criteria
    * @param userid             the id of the user that releases the inquiry
    * @param inquiryName        the label of the inquiry
    * @param inquiryDescription the description of the inquiry
    * @param exposeId           the id of the expose linked with this inquiry
    * @param voteId             the id of the vote linked with this inquiry
    * @param resultTypes        list of the entities that are searched for
-   * @param isMonitoring       if true, no creation of cql query
    * @return the id of the inquiry draft
    */
-  private int store(String query, int userid, String inquiryName,
-      String inquiryDescription, int exposeId, int voteId, List<String> resultTypes,
-      boolean isMonitoring) {
+  private int store(InquiryCriteriaTranslatable query, int userid, String inquiryName,
+      String inquiryDescription, int exposeId, int voteId, List<String> resultTypes)
+      throws IllegalStateException {
     int returnValue = 0;
     UserDao userDao;
     User user;
@@ -204,24 +166,10 @@ public class InquiryHandler {
       inquiry.setResultType(joinedResultTypes);
 
       Record inquiryRecord = saveInquiry(inquiry, connection);
-
       inquiry.setCreated(inquiryRecord.getValue(Tables.INQUIRY.CREATED));
       inquiry.setId(inquiryRecord.getValue(Tables.INQUIRY.ID));
-
-      if (!isMonitoring) {
-        EssentialSimpleQueryDto essentialSimpleQueryDto = jsonString2EssentialDto(query);
-        createAndSaveInquiryCriteria(essentialSimpleQueryDto, inquiry, connection);
-        createAndSaveStatistics(essentialSimpleQueryDto, inquiry.getId());
-      } else {
-        CqlQueryList cqlQueryList = gson.fromJson(query, CqlQueryList.class);
-        for (CqlQuery cqlQuery : cqlQueryList.getQueries()) {
-          InquiryCriteria inquiryCriteria = new InquiryCriteria();
-          inquiryCriteria.setCriteria(cqlQuery.getCql());
-          inquiryCriteria.setInquiryId(inquiry.getId());
-          inquiryCriteria.setType(InquiryCriteriaType.IC_CQL);
-          inquiryCriteria.setEntityType(cqlQuery.getEntityType());
-          saveInquiryCriteria(inquiryCriteria, connection);
-        }
+      for (InquiryCriteria inquiryCriteria : query.createCriteria(inquiry)) {
+        saveInquiryCriteria(inquiryCriteria, connection);
       }
       if (exposeId > 0) {
         Document expose = DocumentUtil.getDocumentById(exposeId);
@@ -243,12 +191,6 @@ public class InquiryHandler {
     }
 
     return returnValue;
-  }
-
-  private void createAndSaveStatistics(EssentialSimpleQueryDto essentialSimpleQueryDto,
-      Integer inquiryId) {
-    StatisticsHandler statisticsHandler = new StatisticsHandler();
-    statisticsHandler.save(essentialSimpleQueryDto, inquiryId);
   }
 
   /**
@@ -313,85 +255,6 @@ public class InquiryHandler {
   }
 
   /**
-   * Spawn a project for an edited inquiry. Or change the status if needed.
-   *
-   * @param inquiry     the inquiry
-   * @param resultTypes a list of expected result types from the inquiry
-   * @param userid      the id of the user that created the inquiry
-   */
-  public void spawnProjectIfNeeded(Inquiry inquiry, List<String> resultTypes, int userid) {
-    UserDao userDao;
-    User user;
-    InquiryDao inquiryDao;
-
-    String projectName = ProjectInfo.INSTANCE.getProjectName();
-    boolean isDktk = projectName.equalsIgnoreCase("dktk");
-
-    if (!isDktk) {
-      logger.debug("Not dktk...no project will be spawned");
-      return;
-    } else if (inquiry.getProjectId() != null) {
-      logger.debug("Project is already spawned...change status");
-      Project project = ProjectUtil.fetchProjectById(inquiry.getProjectId());
-      project.setStatus(ProjectStatus.PS_OPEN_MOREINFO_PROVIDED);
-      project.setSeen(false);
-      ProjectUtil.updateProject(project);
-
-      Action action = new Action();
-      action.setProjectId(project.getId());
-      action.setMessage(ActionType.AT_INQUIRY_EDITED.toString());
-      action.setType(ActionType.AT_PROJECT_CALLBACK_RECEIVED);
-      action.setTime(SamplyShareUtils.getCurrentTime());
-      action.setUserId(userid);
-      ActionUtil.insertAction(action);
-      return;
-    }
-
-    logger.debug("Spawning project");
-
-    try (Connection connection = ResourceManager.getConnection()) {
-      Configuration configuration = new DefaultConfiguration().set(connection)
-          .set(SQLDialect.POSTGRES);
-
-      userDao = new UserDao(configuration);
-      user = userDao.fetchOneById(userid);
-      if (user == null) {
-        return;
-      }
-
-      inquiry.setStatus(InquiryStatus.IS_RELEASED);
-
-      String joinedResultTypes = "";
-      try {
-        Joiner joiner = Joiner.on(",");
-        joinedResultTypes = joiner.join(resultTypes);
-      } catch (Exception e) {
-        logger.warn("couldn't join result types: " + e);
-
-      }
-      inquiry.setResultType(joinedResultTypes);
-      DSLContext dslContext = ResourceManager.getDslContext(connection);
-      // Use the inquiry name as a preliminary project name. Has to be
-      // changed later when it becomes a project.
-      Record record = dslContext
-          .insertInto(Tables.PROJECT, Tables.PROJECT.PROJECTLEADER_ID, Tables.PROJECT.STATUS,
-              Tables.PROJECT.NAME)
-          .values(inquiry.getAuthorId(), ProjectStatus.PS_NEW, inquiry.getLabel())
-          .returning(Tables.PROJECT.ID, Tables.PROJECT.APPLICATION_NUMBER).fetchOne();
-
-      inquiryDao = new InquiryDao(configuration);
-      // Reload the inquiry to fill the expires field
-      int projectId = record.getValue(Tables.PROJECT.ID);
-      inquiry.setProjectId(projectId);
-      inquiryDao.update(inquiry);
-      int inquiryId = inquiry.getId();
-      DocumentUtil.setProjectIdForDocumentByInquiryId(inquiryId, projectId);
-    } catch (SQLException e) {
-      logger.debug("Sql exception caught: " + e);
-    }
-  }
-
-  /**
    * Creates a new tentative inquiry. Tentative inquiries are those that are submitted by the
    * central search tool. They will be purged after a short period of time if no action by the user
    * is taken.
@@ -401,40 +264,33 @@ public class InquiryHandler {
    * @return the id of the newly created tentative inquiry
    */
   int createTentative(Query query, int userid) {
-    int returnValue = 0;
-    UserDao userDao;
-    User user;
-    Inquiry inquiry;
-
     try (Connection connection = ResourceManager.getConnection()) {
       Configuration configuration = new DefaultConfiguration().set(connection)
           .set(SQLDialect.POSTGRES);
 
-      userDao = new UserDao(configuration);
-      user = userDao.fetchOneById(userid);
+      UserDao userDao = new UserDao(configuration);
+      User user = userDao.fetchOneById(userid);
       if (user == null) {
         return 0;
       }
 
-      inquiry = new Inquiry();
+      Inquiry inquiry = new Inquiry();
       inquiry.setAuthorId(user.getId());
 
       // Revision 0 = tentative
       inquiry.setRevision(0);
-
       inquiry.setStatus(InquiryStatus.IS_DRAFT);
 
-      returnValue = saveTentativeInquiry(inquiry, connection).getValue(Tables.INQUIRY.ID);
+      int inquiryId = saveTentativeInquiry(inquiry, connection).getValue(Tables.INQUIRY.ID);
 
-      createAndSaveInquiryCriteriaTypeQuery(query, inquiry, connection);
-    } catch (JAXBException e1) {
-      e1.printStackTrace();
-      return 0;
+      DefaultInquiryCriteriaTranslatable.createInquiryCriteriaTypeQuery(query, inquiry)
+          .ifPresent(criteria -> saveInquiryCriteria(criteria, connection));
+
+      return inquiryId;
     } catch (SQLException e) {
-      e.printStackTrace();
+      logger.error(e);
+      return 0;
     }
-
-    return returnValue;
   }
 
   /**
@@ -482,8 +338,6 @@ public class InquiryHandler {
    * @return true on success, false on error
    */
   public boolean setSitesForInquiry(int inquiryId, List<String> siteIds) {
-    boolean ret = true;
-
     try (Connection connection = ResourceManager.getConnection()) {
       DSLContext dslContext = ResourceManager.getDslContext(connection);
 
@@ -498,12 +352,11 @@ public class InquiryHandler {
                 Tables.INQUIRY_SITE.SITE_ID)
             .values(inquiryId, Integer.parseInt(site)).execute();
       }
-
+      return true;
     } catch (SQLException e) {
-      logger.error("Error adding sites for distribution for inquiry " + inquiryId);
-      ret = false;
+      logger.error("Error adding sites for distribution for inquiry " + inquiryId, e);
+      return false;
     }
-    return ret;
   }
 
   /**
@@ -609,8 +462,8 @@ public class InquiryHandler {
       de.samply.share.model.common.Inquiry inq = new de.samply.share.model.common.Inquiry();
 
       UriBuilder uriBuilder = uriInfo.getBaseUriBuilder();
-      URI uri = uriBuilder.build();
-      inq.setExposeUrl(uri + "searchbroker/exposes/" + inquiryId);
+      inq.setExposeUrl(
+          uriBuilder.path(Searchbroker.class, "getSynopsis").build(inquiryId).toString());
       inq.setId(Integer.toString(inquiryId));
       if (inquiry.getRevision() != null && inquiry.getRevision() > 0) {
         inq.setRevision(Integer.toString(inquiry.getRevision()));
@@ -625,6 +478,8 @@ public class InquiryHandler {
         addCriteriaForViewQuery(inquiryId, inq);
       } else if (isQueryLanguageCql(queryLanguage)) {
         addCriteriaForCql(inquiryId, inq);
+      } else if (isQueryLanguageStructuredQuery(queryLanguage)) {
+        addCriteriaForStructuredQuery(inquiryId, inq);
       } else {
         logger.warn(
             "Header parameter '" + Constants.HEADER_KEY_QUERY_LANGUAGE + "' has invalid value '"
@@ -648,13 +503,13 @@ public class InquiryHandler {
       StringWriter stringWriter = new StringWriter();
       JAXBContext jaxbContext = JAXBContext.newInstance(ObjectFactory.class);
       Marshaller marshaller = jaxbContext.createMarshaller();
-      marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
+      marshaller.setProperty(JAXB_FRAGMENT, TRUE);
 
       ObjectFactory objectFactory = new ObjectFactory();
 
       marshaller.marshal(objectFactory.createInquiry(inq), stringWriter);
 
-      returnValue.append(stringWriter.toString());
+      returnValue.append(stringWriter);
 
     } catch (SQLException e) {
       e.printStackTrace();
@@ -692,6 +547,13 @@ public class InquiryHandler {
     }
 
     inq.setCqlQueryList(cqlQueryList);
+  }
+
+  private void addCriteriaForStructuredQuery(int inquiryId,
+      de.samply.share.model.common.Inquiry inq) {
+    String criteria = InquiryCriteriaUtil.fetchCriteriaListForInquiryIdTypeStructuredQuery(
+        inquiryId);
+    inq.setStructuredQuery(criteria);
   }
 
   /**
@@ -770,7 +632,7 @@ public class InquiryHandler {
       ObjectFactory objectFactory = new ObjectFactory();
       StringWriter stringWriter = new StringWriter();
       Marshaller marshaller = jaxbContext.createMarshaller();
-      marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
+      marshaller.setProperty(JAXB_FRAGMENT, TRUE);
       marshaller.marshal(objectFactory.createContact(contact), stringWriter);
 
       ret = stringWriter.toString();
@@ -852,17 +714,13 @@ public class InquiryHandler {
    * @return the description
    */
   String getInfo(int inquiryId) throws JAXBException {
-    Inquiry inquiry;
-    InquiryDao inquiryDao;
-
-    String ret = "";
 
     try (Connection connection = ResourceManager.getConnection()) {
       Configuration configuration = new DefaultConfiguration().set(connection)
           .set(SQLDialect.POSTGRES);
 
-      inquiryDao = new InquiryDao(configuration);
-      inquiry = inquiryDao.findById(inquiryId);
+      InquiryDao inquiryDao = new InquiryDao(configuration);
+      Inquiry inquiry = inquiryDao.findById(inquiryId);
 
       Info info = new Info();
 
@@ -883,103 +741,15 @@ public class InquiryHandler {
       ObjectFactory objectFactory = new ObjectFactory();
       StringWriter stringWriter = new StringWriter();
       Marshaller marshaller = jaxbContext.createMarshaller();
-      marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
+      marshaller.setProperty(JAXB_FRAGMENT, TRUE);
       marshaller.marshal(objectFactory.createInfo(info), stringWriter);
 
-      ret = stringWriter.toString();
+      return stringWriter.toString();
 
     } catch (SQLException e) {
-      e.printStackTrace();
+      logger.error(e);
+      return "";
     }
-    return ret;
-  }
-
-  private void createAndSaveInquiryCriteria(EssentialSimpleQueryDto essentialSimpleQueryDto,
-      Inquiry inquiry, Connection connection) {
-    createAndSaveInquiryCriteriaTypeCql(essentialSimpleQueryDto, inquiry, connection);
-    createAndSaveInquiryCriteriaTypeQuery(essentialSimpleQueryDto, inquiry, connection);
-  }
-
-  private void createAndSaveInquiryCriteriaTypeCql(EssentialSimpleQueryDto essentialSimpleQueryDto,
-      Inquiry inquiry, Connection connection) {
-    createAndSaveInquiryCriteriaTypeCqlPatient(essentialSimpleQueryDto, inquiry, connection);
-    createAndSaveInquiryCriteriaTypeCqlSpecimen(essentialSimpleQueryDto, inquiry, connection);
-  }
-
-  private void createAndSaveInquiryCriteriaTypeCql(String cql, Inquiry inquiry,
-      Connection connection, String entityType) {
-    InquiryCriteria inquiryCriteria = new InquiryCriteria();
-    inquiryCriteria.setCriteria(cql);
-    inquiryCriteria.setInquiryId(inquiry.getId());
-    inquiryCriteria.setType(InquiryCriteriaType.IC_CQL);
-    inquiryCriteria.setEntityType(entityType);
-
-    saveInquiryCriteria(inquiryCriteria, connection);
-  }
-
-  private void createAndSaveInquiryCriteriaTypeCqlPatient(
-      EssentialSimpleQueryDto essentialSimpleQueryDto, Inquiry inquiry, Connection connection) {
-    String cql = createCqlPatient(essentialSimpleQueryDto);
-
-    createAndSaveInquiryCriteriaTypeCql(cql, inquiry, connection, ENTITY_TYPE_FOR_CQL_PATIENT);
-  }
-
-  private void createAndSaveInquiryCriteriaTypeCqlSpecimen(
-      EssentialSimpleQueryDto essentialSimpleQueryDto, Inquiry inquiry, Connection connection) {
-    String cql = createCqlSpecimen(essentialSimpleQueryDto);
-
-    createAndSaveInquiryCriteriaTypeCql(cql, inquiry, connection, ENTITY_TYPE_FOR_CQL_SPECIMEN);
-  }
-
-  private String createCqlPatient(EssentialSimpleQueryDto essentialSimpleQueryDto) {
-    return createCql(essentialSimpleQueryDto, ENTITY_TYPE_FOR_CQL_PATIENT);
-  }
-
-  private String createCqlSpecimen(EssentialSimpleQueryDto essentialSimpleQueryDto) {
-    return createCql(essentialSimpleQueryDto, ENTITY_TYPE_FOR_CQL_SPECIMEN);
-  }
-
-  private String createCql(EssentialSimpleQueryDto essentialSimpleQueryDto, String entityType) {
-    return new EssentialSimpleQueryDto2CqlTransformer()
-        .toQuery(essentialSimpleQueryDto, entityType);
-  }
-
-  private void createAndSaveInquiryCriteriaTypeQuery(
-      EssentialSimpleQueryDto essentialSimpleQueryDto, Inquiry inquiry, Connection connection) {
-    Query query;
-
-    if (CollectionUtils.isEmpty(essentialSimpleQueryDto.getFieldDtos())) {
-      query = new EssentialSimpleQueryDto2ShareXmlTransformer().toQuery(essentialSimpleQueryDto);
-    } else {
-      query = new Query();
-      Where where = new Where();
-      And and = new And();
-      where.getAndOrEqOrLike().add(and);
-      query.setWhere(where);
-    }
-
-    try {
-      createAndSaveInquiryCriteriaTypeQuery(query, inquiry, connection);
-    } catch (JAXBException e) {
-      e.printStackTrace();
-    }
-  }
-
-  private void createAndSaveInquiryCriteriaTypeQuery(Query query, Inquiry inquiry,
-      Connection connection) throws JAXBException {
-    JAXBContext jaxbContext = JAXBContext.newInstance(ObjectFactory.class);
-    StringWriter stringWriter = new StringWriter();
-    Marshaller marshaller = jaxbContext.createMarshaller();
-    marshaller.setProperty(Marshaller.JAXB_FRAGMENT, Boolean.TRUE);
-    marshaller.marshal(query, stringWriter);
-
-    InquiryCriteria inquiryCriteria = new InquiryCriteria();
-    inquiryCriteria.setCriteria(stringWriter.toString());
-    inquiryCriteria.setInquiryId(inquiry.getId());
-    inquiryCriteria.setType(InquiryCriteriaType.IC_QUERY);
-    inquiryCriteria.setEntityType(ENTITY_TYPE_FOR_QUERY);
-
-    saveInquiryCriteria(inquiryCriteria, connection);
   }
 
   private void saveInquiryCriteria(InquiryCriteria inquiryCriteria, Connection connection) {
@@ -1046,9 +816,6 @@ public class InquiryHandler {
    * @return true, if successful
    */
   boolean saveReply(int inquiryId, int bankId, String content, Timestamp timestamp) {
-    boolean returnValue = true;
-    Reply reply;
-    ReplyDao replyDao;
     Gson gson = new GsonBuilder().disableHtmlEscaping().create();
     de.samply.share.model.common.result.Reply replyJson = gson.fromJson(content,
         de.samply.share.model.common.result.Reply.class);
@@ -1060,12 +827,13 @@ public class InquiryHandler {
       Configuration configuration = new DefaultConfiguration().set(connection)
           .set(SQLDialect.POSTGRES);
       DSLContext dslContext = ResourceManager.getDslContext(connection);
-      replyDao = new ReplyDao(configuration);
+      ReplyDao replyDao = new ReplyDao(configuration);
 
       Record record = dslContext.select().from(Tables.REPLY)
           .where(Tables.REPLY.BANK_ID.equal(bankId).and(Tables.REPLY.INQUIRY_ID.equal(inquiryId)))
           .fetchAny();
 
+      Reply reply;
       if (record == null) {
         reply = new Reply();
         reply.setBankId(bankId);
@@ -1079,27 +847,10 @@ public class InquiryHandler {
         reply.setRetrievedat(timestamp);
         replyDao.update(reply);
       }
-
+      return true;
     } catch (SQLException e) {
-      e.printStackTrace();
-      returnValue = false;
+      logger.error(e);
+      return false;
     }
-
-    return returnValue;
   }
-
-  private boolean isEmpty(EssentialSimpleValueDto valueDto) {
-    return isEmptyValue(valueDto.getValue())
-        || (valueDto.getCondition() == SimpleValueCondition.BETWEEN && isEmptyValue(
-        valueDto.getMaxValue()));
-  }
-
-  private boolean isEmpty(EssentialSimpleFieldDto fieldDto) {
-    return CollectionUtils.isEmpty(fieldDto.getValueDtos());
-  }
-
-  private boolean isEmptyValue(String value) {
-    return StringUtils.isEmpty(value) || "null".equalsIgnoreCase(value);
-  }
-
 }